Fix small bug and refactoring
