--- requirements.txt
@@ -1,19 +1,0 @@
-aniso8601==9.0.1
-black==22.1.0
-click==8.0.4
-Flask==2.0.3
-Flask-RESTful==0.3.9
-itsdangerous==2.1.0
-Jinja2==3.0.3
-MarkupSafe==2.1.0
-mypy-extensions==0.4.3
-mysql-connector-python==8.0.28
-pathspec==0.9.0
-platformdirs==2.5.1
-protobuf==3.19.4
-python-decouple==3.6
-pytz==2021.3
-six==1.16.0
-SQLAlchemy==1.4.31
-tomli==2.0.1
-Werkzeug==2.0.3
